backend/requirements.txt
fastapi
uvicorn
python-multipart
pypdf
faiss-cpu
numpy
sentence-transformers
torch